Added simple button class and use it in the StopwatchApp.  Also, fix a warining in TimerApp.

+++ TimerApp.cpp
@@ -27,7 +27,7 @@ BritepadApp* TimerApp::run(void) {
   if (draw) {
     char textTime[6];
 
-    long displaytime = alarm_sounded ? timer_dur : delta;
+    int displaytime = alarm_sounded ? timer_dur : delta;
     bool drawtext = true;
     bool erasescreen = false;
